many changes including support for v5 and v6 versions of oncoplex.

+++ process_variants.py
@@ -11,43 +11,79 @@ genes = list(df['gene'])
 genes = dict(zip(genes,genes))
 genesX = list(df[(df['chrom'] == 'chrX')]['gene'])
 genesX = dict(zip(genesX,genesX))
+chroms = ['chr1','chr2','chr3','chr4','chr5','chr6','chr7','chr8','chr9','chr10','chr11','chr12','chr13','chr14','chr15','chr16','chr17','chr18','chr19','chr20','chr21','chr22','chrX','chrMT','chrM','chrY']
 
 # *** process ***
-def process(inFile,outFile):
+def process(inFile,outFile,pipeline,overwrite = True):
 
-    # ** mutation processing here ** 
+    # check if file exists
+    if os.path.exists(outFile) and not overwrite:
+        print(' ... skipped',end='')
+        return
+    
+#    pipeline = 'TGC2' if 'v7' in inFile or 'v8' in inFile else 'TGC'
+    
+    # ** mutation processing here **
     if 'mutation' in inFile:
         # create output directory if one is not already 
         outDir = os.path.dirname(outFile)
         if not os.path.exists(outDir):
             os.makedirs(outDir, exist_ok=True)
 
+        consequences = ["coding", "frameshift", "inframe", "missense", "splicing-canonical", "start/stop",'frameshift deletion','frameshift insertion','nonframeshift deletion','nonsynonymous SNV','splicing','nonframeshift insertion']
+
+            
         # read variant file
         df1 = pd.read_csv(inFile,sep="\t",keep_default_na = False)
-
-#        print(set(list(df1['sheet'])))
-        
         # ** filtering criteria **
-        idx0 = (df1['sheet'] == 'smallVariants')
-        idx1 = (df1['gnomadFreq'] == 'NA') | (pd.to_numeric(df1['gnomadFreq'],errors='coerce') < 0.001)
-        idx2 = (df1['uwFreq'] < 0.05)  # internal frequency database
-        consequences = ["coding", "frameshift", "inframe", "missense", "splicing-canonical", "start/stop"]    
-        idx3 = (df1['consequence'].isin(consequences))
-        idx4 = (~df1['clinvar'].str.contains('benign',case = False))
-        idx5 = (df1['VAF'] > 0.03)
-        idx6 = (df1['filterRealVariant'] == True)
-        idx10 = (df1['AD'] > 12)  
-        # **
 
-        idx7 = (df1['sheet'] == 'clinicallyFlagged') & (df1['clinvar'].str.contains('pathogenic',case=False)) & (~df1['clinvar'].str.contains('conflicting',case=False))
-        idx8 = (df1['AD'] > 12)  # was 5 based on macro filters
-        idx9 = (df1['uwFreq'] < 0.1)  # based on most frequent mutations in cancer ~5%
-        df1a = df1[(idx0 & idx1 & idx2 & idx3 & idx4 & idx5 & idx6 & idx10) | (idx5 & idx7 & idx8 & idx9)].copy()
+        if pipeline == 'TGC2':
+            idx0 = (df1['sheet'] == 'smallVariants')
+
+            idx1 = (df1['humanFreq'] == 'NA') | (pd.to_numeric(df1['humanFreq'],errors='coerce') < 0.001)
+
+            idx2 = (pd.to_numeric(df1['uwFreq'],errors='coerce') < 0.05)  # intwernal frequency database        
+
+            idx3 = (df1['consequence'].isin(consequences))
+            idx4 = (~df1['clinvar'].str.contains('benign',case = False))
+            idx5 = (pd.to_numeric(df1['VAF'],errors='coerce') > 0.03)
+            idx6 = (df1['filterRealVariant'] == True)
+            idx10 = (pd.to_numeric(df1['AD'],errors='coerce') > 12)  
+            # **
+        
+            idx7 = (df1['sheet'] == 'clinicallyFlagged') & (df1['clinvar'].str.contains('pathogenic',case=False)) & (~df1['clinvar'].str.contains('conflicting',case=False))
+            idx8 = (pd.to_numeric(df1['AD'],errors='coerce') > 12)  # was 5 based on macro filters
+            idx9 = (pd.to_numeric(df1['uwFreq'],errors='coerce') < 0.1)  # based on most frequent mutations in cancer ~5%
+            df1a = df1[(idx0 & idx1 & idx2 & idx3 & idx4 & idx5 & idx6 & idx10) | (idx5 & idx7 & idx8 & idx9)].copy()
     
-        # get rid of duplicates from small variants and clinically flagged
-        df1b = df1a.drop_duplicates(subset=['id'],keep = 'first')
-        df1b.to_csv(outFile,index = False, sep="\t")
+            # get rid of duplicates from small variants and clinically flagged
+            df1b = df1a.drop_duplicates(subset=['id'],keep = 'first')
+            df1b = df1b.replace('','NA')
+            df1b.to_csv(outFile,index = False, sep="\t")
+        # different filtering for older pipelines
+        elif pipeline == 'TGC':
+            idx0 = (df1['sheet'] == 'snpAnalysisTxt')
+            idx1 = (df1['humanFreq'] == 'NA') | (pd.to_numeric(df1['humanFreq'],errors='coerce') < 0.001)
 
+            idx2 = (df1['uwFreq'] == 'NA') | (pd.to_numeric(df1['uwFreq'],errors='coerce') < 0.05)  # internal frequency database
+
+            # look for matches to consequence list
+            regex = '|'.join(map(re.escape,consequences))
+            idx3 = (df1['consequence'].str.contains(regex,na=False))
+
+            idx4 = (~df1['clinvar'].str.contains('benign',case = False))
+            idx5 = (df1['VAF'] == 'NA') | (pd.to_numeric(df1['VAF'],errors='coerce') > 0.03)
+            idx10 = (pd.to_numeric(df1['AD'],errors='coerce') > 12)  
+
+            df1a = df1[(idx0 & idx1 & idx2 & idx3 & idx4 & idx5 & idx10)].copy()
+            
+            # get rid of duplicates from small variants and clinically flagged
+            df1b = df1a.drop_duplicates(subset=['id'],keep = 'first')
+            df1b = df1b.replace('','NA')            
+            df1b.to_csv(outFile,index = False, sep="\t")
+        else:
+            raise 'i no understand mutations'
+            
         return df1b  # for troubleshooting
 
     # ** cnvs **
@@ -85,7 +121,7 @@ def process(inFile,outFile):
             elif value >= 0.6:   
                 call = 'GAIN'
                 gisticValue = 1
-            elif value <= -1.3 or (value <= -1 and absMax > 3):  # focal homodels
+            elif value <= -1.3 or (pipeline == 'TGC2' and value <= -1 and absMax > 3):  # focal homodels
                 call = 'HOMODEL'
                 gisticValue = -2
             elif value <= -0.4:
@@ -112,9 +148,8 @@ def process(inFile,outFile):
             else:
                 umbrellaCall = 'intact'
             df1.at[idx,'umbrellaID'] = '__'.join([sample,gene,umbrellaCall])                 
-
         df1a = df1[['id','umbrellaID','sample','gene','avgLogRatio','maxAbsLogRatio','call','gisticValue','del3ExonValue','gain3ExonValue']]
-        df1a.to_csv(outFile,index = False, sep='\t')
+        df1a.to_csv(outFile,index = False, sep='\t',na_rep = 'NA')
     # **
 
     # ** svs **
@@ -127,16 +162,26 @@ def process(inFile,outFile):
         # read variant file
         df1 = pd.read_csv(inFile,sep="\t",keep_default_na = False,na_values = ['NA','nan'])
         df1['svLength'] = -1
-            
-        # ** filtering criteria **
-        idx1 = (df1['VAF'] > 0.008)
-        idx2 = (df1['score'] > 1500) | ((df1['score'] > 400) & (df1['isQuiver'] | df1['isMitelman']))
-        idx3 = (df1['filterRealVariant'] == True)
-        # group 0 are consider more likely to be FP by dlmp
-        idx4 = ((df1['group'] == 0) & (df1['isQuiver'] | df1['isMitelman'] | ~df1['id'].str.contains('Intergenic'))) | (df1['group'] != 0)
-        df1a = df1[idx1 & idx2 & idx3 & idx4].copy()
 
-        df1a = df1a.reset_index(drop = True)
+        if pipeline == 'TGC2':
+            # ** filtering criteria **
+            idx1 = (df1['VAF'] > 0.008)
+            idx2 = (df1['score'] > 1500) | ((df1['score'] > 400) & (df1['isQuiver'] | df1['isMitelman']))
+            idx3 = (df1['filterRealVariant'] == True)
+            # group 0 are consider more likely to be FP by dlmp
+            idx4 = ((df1['group'] == 0) & (df1['isQuiver'] | df1['isMitelman'] | ~df1['id'].str.contains('Intergenic'))) | (df1['group'] != 0)
+            df1a = df1[idx1 & idx2 & idx3 & idx4].copy()
+
+            df1a = df1a.reset_index(drop = True)
+
+        # **
+        elif pipeline == 'TGC':
+            idx1 = (df1['chrom1'].isin(chroms)) & (df1['chrom2'].isin(chroms))
+            idx2 = (df1['gene1'].isin(genes)) | (df1['gene2'].isin(genes))
+            idx3 = (df1['gene1'] != df1['gene2'])
+
+            df1a = df1[idx1 & idx2 & idx3].copy()
+            df1a = df1a.reset_index(drop = True)
 
         # ** add svLength **
         if len(df1a) > 0:        
@@ -153,11 +198,12 @@ def process(inFile,outFile):
                 else:
                     svLength = float(pos1) - float(pos2) + 1
                 df1a.at[idx,'svLength'] = -1 if chrom1 != chrom2 else svLength
-        # **
+        # ** 
         
         # get rid of duplicates
         df1b = df1a.drop_duplicates(subset=['id'],keep = 'first')
-        df1b.to_csv(outFile,index = False, sep="\t")
+        df1b = df1b.replace('','NA')                    
+        df1b.to_csv(outFile,index = False, sep="\t",na_rep = 'NA')
             
         return df1b # for troubleshooting
     # **
